refactor : divide ui and logic at ProfileScreen #92
